Revert ":recycle; Rename package to fetcher"
This reverts commit bcad722f41e433819b25e3fbf534cc28f2e608c7.
