+++ pkg/parser/size.go
@@ -1,4 +1,4 @@
-package fetcher
+package parser
 
 import (
 	"fmt"
